requirements.txt
pdfplumber==0.7.4
requests==2.26.0
datasets==2.3.2
transformers==4.18.0
langchain==0.0.184
python-dotenv==0.19.2
tqdm==4.62.3